requirements.txt
pymupdf 
streamlit 
google-generativeai